Add files via upload
The project was a part of "Data Visualization And Interpretation". The task involved the identification of the most effective visualization techniques, depending upon the dependent and independent features, for univariate and multivariate analysis. The work included the use of eight Gestalt principles of visual perception, decluttering techniques, and twelve pre-attentive attributes for effective visualization. These techniques help in directing the audience's attention to the most important take-away from the data.